fix: Rename the net5app sample and other referneces as dotnet5
