+++ samples/dotnet5/.dockerignore
