+++ samples/dotnet5/Pages/Error.cshtml
