+++ samples/dotnet5/Pages/Privacy.cshtml
