+++ samples/dotnet5/Pages/Shared/_ValidationScriptsPartial.cshtml
