+++ samples/dotnet5/Pages/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-@using Net5app
-@namespace Net5app.Pages
+@using DotNet5app
+@namespace DotNet5app.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
