+++ samples/dotnet5/Pages/_ViewStart.cshtml
