+++ samples/dotnet5/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace Net5app
+namespace DotNet5app
 {
     public class Program
     {
