+++ samples/dotnet5/Startup.cs
@@ -9,7 +9,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Net5app
+namespace DotNet5app
 {
     public class Startup
     {
