+++ samples/dotnet5/appsettings.Development.json
