+++ samples/dotnet5/dotnet5.csproj
