+++ samples/dotnet5/dotnet5.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.809.10
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "net5app", "net5app.csproj", "{984A45C3-5D43-440F-953A-3A0AD18EF503}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnet5", "dotnet5.csproj", "{984A45C3-5D43-440F-953A-3A0AD18EF503}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
