+++ samples/dotnet5/wwwroot/favicon.ico
