+++ samples/dotnet5/wwwroot/js/site.js
