+++ samples/dotnet5/wwwroot/lib/bootstrap/LICENSE
