+++ samples/dotnet5/wwwroot/lib/jquery-validation-unobtrusive/LICENSE.txt
