+++ samples/dotnet5/wwwroot/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js
