+++ samples/dotnet5/wwwroot/lib/jquery-validation/LICENSE.md
