+++ samples/dotnet5/wwwroot/lib/jquery/LICENSE.txt
